Split settings toggle component to its own files
Fix #7153

Auditors: @bsclifton
