+++ app/renderer/components/settings.js
@@ -4,6 +4,10 @@
 
 const React = require('react')
 const ImmutableComponent = require('../../../js/components/immutableComponent')
+const aboutActions = require('../../../js/about/aboutActions')
+const getSetting = require('../../../js/settings').getSetting
+const {changeSetting} = require('../lib/settingsUtil')
+const SwitchControl = require('../../../js/components/switchControl')
 
 class SettingsList extends ImmutableComponent {
   render () {
@@ -33,7 +37,72 @@ class SettingItem extends ImmutableComponent {
   }
 }
 
+class SettingCheckbox extends ImmutableComponent {
+  constructor () {
+    super()
+    this.onClick = this.onClick.bind(this)
+  }
+
+  onClick (e) {
+    if (this.props.disabled) {
+      return
+    }
+    return this.props.onChange ? this.props.onChange(e) : changeSetting(this.props.onChangeSetting, this.props.prefKey, e)
+  }
+
+  render () {
+    const props = {
+      style: this.props.style,
+      className: 'settingItem'
+    }
+    if (this.props.id) {
+      props.id = this.props.id
+    }
+    return <div {...props}>
+      <SwitchControl id={this.props.prefKey}
+        small={this.props.small}
+        disabled={this.props.disabled}
+        onClick={this.onClick}
+        checkedOn={this.props.checked !== undefined ? this.props.checked : getSetting(this.props.prefKey, this.props.settings)} />
+      <label data-l10n-id={this.props.dataL10nId} htmlFor={this.props.prefKey} />
+      {this.props.options}
+    </div>
+  }
+}
+
+class SiteSettingCheckbox extends ImmutableComponent {
+  constructor () {
+    super()
+    this.onClick = this.onClick.bind(this)
+  }
+
+  onClick (e) {
+    if (this.props.disabled || !this.props.hostPattern) {
+      return
+    } else {
+      const value = !!e.target.value
+      value === this.props.defaultValue
+        ? aboutActions.removeSiteSetting(this.props.hostPattern,
+            this.props.prefKey)
+        : aboutActions.changeSiteSetting(this.props.hostPattern,
+            this.props.prefKey, value)
+    }
+  }
+
+  render () {
+    return <div style={this.props.style} className='settingItem siteSettingItem'>
+      <SwitchControl
+        small={this.props.small}
+        disabled={this.props.disabled}
+        onClick={this.onClick}
+        checkedOn={this.props.checked} />
+    </div>
+  }
+}
+
 module.exports = {
   SettingsList,
-  SettingItem
+  SettingItem,
+  SettingCheckbox,
+  SiteSettingCheckbox
 }
